requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# LINE API
line-bot-sdk==3.5.0

# HTTP Client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
loguru==0.7.2 